simplified multisig interface helper
