+++ lib/multiSig-tasks/confirm-proposal.ts
@@ -2,12 +2,7 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_CONFIRM_PROPOSAL } from "../tasksNames";
 
-import {
-  getSigner,
-  getMultiSig,
-  // confirmProposal,
-  parseEvents,
-} from "../helpers/multiSigInterfaceHelper";
+import { getSigner, parseEvents } from "../helpers/interfaceHelper";
 
 task(MULTISIG_CONFIRM_PROPOSAL, "Confirm a multiSig proposal")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
@@ -16,7 +11,7 @@ task(MULTISIG_CONFIRM_PROPOSAL, "Confirm a multiSig proposal")
   .setAction(async ({ proposalId, namedAccount, useLedger }, hre) => {
     try {
       const signer = await getSigner(hre, namedAccount, useLedger);
-      const multiSigContract = await getMultiSig(hre);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
       const tx = await multiSigContract.connect(signer).confirmProposal(proposalId);
       const receipt = await tx.wait();
       parseEvents(receipt, "ProposalConfirmed");
