+++ lib/multiSig-tasks/get-confirmations.ts
@@ -2,14 +2,12 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_GET_CONFIRMATIONS } from "../tasksNames";
 
-import { getMultiSig, getConfirmations } from "../helpers/multiSigInterfaceHelper";
-
 task(MULTISIG_GET_CONFIRMATIONS, "Get proposal confirmations")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
   .setAction(async ({ proposalId }, hre) => {
     try {
-      const multiSigContract = await getMultiSig(hre);
-      const confirmations = await getConfirmations(multiSigContract, proposalId);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const confirmations = await multiSigContract.getConfirmations(proposalId);
       console.log(confirmations);
     } catch (error) {
       console.log(error);
