+++ lib/multiSig-tasks/get-owners.ts
@@ -2,12 +2,10 @@ import { task } from "hardhat/config";
 
 import { MULTISIG_GET_OWNERS } from "../tasksNames";
 
-import { getMultiSig, getOwners } from "../helpers/multiSigInterfaceHelper";
-
 task(MULTISIG_GET_OWNERS, "Get multiSig owners").setAction(async (_, hre) => {
   try {
-    const multiSigContract = await getMultiSig(hre);
-    const owners = await getOwners(multiSigContract);
+    const multiSigContract = await hre.ethers.getContract("MultiSig");
+    const owners = await multiSigContract.getOwners();
     console.log(owners);
   } catch (error) {
     console.log(error);
