+++ lib/multiSig-tasks/is-confirmed-by.ts
@@ -2,15 +2,13 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_IS_CONFIRMED_BY } from "../tasksNames";
 
-import { getMultiSig, isConfirmedBy } from "../helpers/multiSigInterfaceHelper";
-
 task(MULTISIG_IS_CONFIRMED_BY, "Check if a proposal has been confirmed a multiSig owner")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
   .addParam("address", "Owner address", undefined, types.string)
   .setAction(async ({ proposalId, address }, hre) => {
     try {
-      const multiSigContract = await getMultiSig(hre);
-      const result = await isConfirmedBy(multiSigContract, proposalId, address);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const result = await multiSigContract.isConfirmedBy(proposalId, address);
       console.log(result);
     } catch (error) {
       console.log(error);
