+++ lib/multiSig-tasks/is-fully-confirmed.ts
@@ -2,14 +2,12 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_IS_FULLY_CONFIRMED } from "../tasksNames";
 
-import { getMultiSig, isFullyConfirmed } from "../helpers/multiSigInterfaceHelper";
-
 task(MULTISIG_IS_FULLY_CONFIRMED, "Check if a proposal has been fully confirmed")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
   .setAction(async ({ proposalId }, hre) => {
     try {
-      const multiSigContract = await getMultiSig(hre);
-      const fullyConfirmed = await isFullyConfirmed(multiSigContract, proposalId);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const fullyConfirmed = await multiSigContract.isFullyConfirmed(proposalId);
       console.log(fullyConfirmed);
     } catch (error) {
       console.log(error);
