+++ lib/multiSig-tasks/revoke-confirmation.ts
@@ -2,12 +2,7 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_REVOKE_CONFIRMATION } from "../tasksNames";
 
-import {
-  getSigner,
-  getMultiSig,
-  revokeConfirmation,
-  parseEvents,
-} from "../helpers/multiSigInterfaceHelper";
+import { getSigner, parseEvents } from "../helpers/interfaceHelper";
 
 task(MULTISIG_REVOKE_CONFIRMATION, "Revoke a proposal confirmation")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
@@ -16,8 +11,9 @@ task(MULTISIG_REVOKE_CONFIRMATION, "Revoke a proposal confirmation")
   .setAction(async ({ proposalId, namedAccount, useLedger }, hre) => {
     try {
       const signer = await getSigner(hre, namedAccount, useLedger);
-      const multiSigContract = await getMultiSig(hre);
-      const receipt = await revokeConfirmation(multiSigContract, proposalId, signer);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const tx = await multiSigContract.connect(signer).revokeConfirmation(proposalId);
+      const receipt = await tx.wait();
       parseEvents(receipt, "ConfirmationRevoked");
     } catch (error) {
       console.log(error);
