+++ lib/multiSig-tasks/schedule-proposal.ts
@@ -2,12 +2,7 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_SCHEDULE_PROPOSAL } from "../tasksNames";
 
-import {
-  getSigner,
-  getMultiSig,
-  scheduleProposal,
-  parseEvents,
-} from "../helpers/multiSigInterfaceHelper";
+import { getSigner, parseEvents } from "../helpers/interfaceHelper";
 
 task(MULTISIG_SCHEDULE_PROPOSAL, "Schedule a proposal")
   .addParam("proposalId", "ID of the proposal", undefined, types.int)
@@ -16,8 +11,9 @@ task(MULTISIG_SCHEDULE_PROPOSAL, "Schedule a proposal")
   .setAction(async ({ proposalId, namedAccount, useLedger }, hre) => {
     try {
       const signer = await getSigner(hre, namedAccount, useLedger);
-      const multiSigContract = await getMultiSig(hre);
-      const receipt = await scheduleProposal(multiSigContract, proposalId, signer);
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const tx = await multiSigContract.connect(signer).scheduleProposal(proposalId);
+      const receipt = await tx.wait();
       parseEvents(receipt, "ProposalScheduled");
     } catch (error) {
       console.log(error);
