+++ lib/multiSig-tasks/submit-proposal.ts
@@ -2,7 +2,7 @@ import { task, types } from "hardhat/config";
 
 import { MULTISIG_SUBMIT_PROPOSAL } from "../tasksNames";
 
-import { getMultiSig, getSigner, submitProposal } from "../helpers/multiSigInterfaceHelper";
+import { getSigner } from "../helpers/interfaceHelper";
 
 task(MULTISIG_SUBMIT_PROPOSAL, "Submit a proposal to the multiSig contract")
   .addParam(
@@ -18,14 +18,11 @@ task(MULTISIG_SUBMIT_PROPOSAL, "Submit a proposal to the multiSig contract")
   .setAction(async ({ destinations, values, payloads, namedAccount, useLedger }, hre) => {
     try {
       const signer = await getSigner(hre, namedAccount, useLedger);
-      const multiSigContract = await getMultiSig(hre);
-      const receipt = await submitProposal(
-        multiSigContract,
-        destinations.split(","),
-        values.split(","),
-        payloads.split(","),
-        signer
-      );
+      const multiSigContract = await hre.ethers.getContract("MultiSig");
+      const tx = await multiSigContract
+        .connect(signer)
+        .submitProposal(destinations.split(","), values.split(","), payloads.split(","));
+      const receipt = await tx.wait();
       const events = receipt.events;
       if (events !== undefined) {
         for (var i = 0; i < events!.length; i++) {
